冗長なdir削除しました: download_html_page
